requirements.txt
glfw==2.5.4
imgui==1.4.1
numpy==1.23.1
PyGLM==2.5.7
PyOpenGL==3.1.6
